+++ src/docs/pages/components/form/toggle.vue
@@ -4,15 +4,21 @@
       components : [
         { class: 'form-control', desc: 'Container element' },
         { class: 'toggle', desc: 'For checkbox' },
-        { class: 'toggle-mark', desc: 'For span that coms after checkbox' },
       ],
       utilities : [
         { class: 'toggle-primary', desc: 'Adds `primary` to toggle' },
         { class: 'toggle-secondary', desc: 'Adds `secondary` to toggle' },
         { class: 'toggle-accent', desc: 'Adds `accent` to toggle' },
+        { class: 'toggle-lg', desc: 'Large toggle' },
+        { class: 'toggle-md', desc: 'Medium toggle (default)' },
+        { class: 'toggle-sm', desc: 'Small toggle' },
+        { class: 'toggle-xs', desc: 'Extra small toggle' },
       ]
     }"/>
     <Wrapper title="toggle" classes="flex flex-col space-y-2 max-w-xs">
+      <Toggle/>
+    </Wrapper>
+    <Wrapper title="toggle with label and container" classes="flex flex-col space-y-2 max-w-xs">
       <div class="p-6 card bordered">
         <FormControl>
           <label class="cursor-pointer label">
@@ -52,24 +58,24 @@
         </FormControl>
       </div>
     </Wrapper>
+    <Wrapper title="toggle sizes" classes="flex flex-col space-y-2 max-w-xs">
+      <Toggle classes="toggle-xs" />
+      <Toggle classes="toggle-sm" />
+      <Toggle classes="toggle-md" />
+      <Toggle classes="toggle-lg" />
+    </Wrapper>
     <Wrapper title="Disabled" classes="flex flex-col space-y-2 max-w-xs">
       <div class="p-6 card bordered">
         <FormControl>
           <label class="label">
             <span class="label-text">Unchecked + Disabled</span>
-            <div>
-              <input type="checkbox" class="toggle" :class="classes" disabled="disabled" />
-              <span class="toggle-mark"></span>
-            </div>
+            <input type="checkbox" class="toggle" :class="classes" disabled="disabled" />
           </label>
         </FormControl>
         <FormControl>
           <label class="label">
             <span class="label-text">Checked + Disabled</span>
-            <div>
-              <input type="checkbox" class="toggle" :class="classes" checked disabled="disabled" />
-              <span class="toggle-mark"></span>
-            </div>
+            <input type="checkbox" class="toggle" :class="classes" checked disabled="disabled" />
           </label>
         </FormControl>
       </div>